Fixed project conversion bugs in X10DT.
git-svn-id: svn://svn.code.sf.net/p/x10/code/x10dt/trunk@26372 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10dt.ui.builder.java/src/x10dt/ui/launch/java/builder/X10JavaBuilder.java
@@ -110,10 +110,12 @@ public class X10JavaBuilder extends AbstractX10Builder {
 		final IWorkspaceRoot root = ResourcesPlugin.getWorkspace().getRoot();
 	    final URI outputFolderURI = URIUtils.getExpectedURI(root.getFolder(fProjectWrapper.getPath().append(new Path("x10-gen-src"))).getLocationURI());
 	    File folder = EFS.getStore(outputFolderURI).toLocalFile(EFS.NONE, new NullProgressMonitor());
-	    for (File file: folder.listFiles()){
-	    	if (file.exists()){
-  			  file.delete();
-  		  }
+	    if (folder != null && folder.exists()){
+	    	for (File file: folder.listFiles()){
+	    		if (file.exists()){
+	    			file.delete();
+	    		}
+	    	}
 	    }
 	}
